+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - added `function.{list,delete,exists}`, `class.{list,delete,exists}` super functions
 - added `out_*` magic functions to handle program stdout output in advance level
 - added multiline using `\` character in end of lines
+- added complicated declaring class and function
 
 #### Bug Fixes
 - fixed a bug in `eval` function